+++ user-service/build.gradle
@@ -29,7 +29,7 @@ ext {
 }
 
 dependencies {
-    implementation 'io.github.boeingmerryho:common-library:1.0.4'
+    implementation 'io.github.boeingmerryho:common-library:1.1.2'
 
     // Spring Boot
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
